more tests that cause OOM
