+++ sdk/storage/azblob/blockblob/client_test.go
@@ -29,7 +29,6 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/streaming"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/internal/recording"
-	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/blob"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/bloberror"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/blockblob"
@@ -3793,39 +3792,6 @@ func (s *BlockBlobUnrecordedTestsSuite) TestBlockBlobSetExpiryToPast() {
 	_require.Nil(resp.ExpiresOn)
 }
 
-func (s *BlockBlobUnrecordedTestsSuite) TestLargeBlockBlobStage() {
-	_require := require.New(s.T())
-	testName := s.T().Name()
-	svcClient, err := testcommon.GetServiceClient(s.T(), testcommon.TestAccountDefault, nil)
-	_require.NoError(err)
-
-	containerName := testcommon.GenerateContainerName(testName)
-	containerClient := testcommon.CreateNewContainer(context.Background(), _require, containerName, svcClient)
-	defer testcommon.DeleteContainer(context.Background(), _require, containerClient)
-
-	bbClient := testcommon.GetBlockBlobClient(testcommon.GenerateBlobName(testName), containerClient)
-
-	var largeBlockSize int64 = blockblob.MaxStageBlockBytes
-	content := make([]byte, largeBlockSize)
-	body := bytes.NewReader(content)
-	rsc := streaming.NopCloser(body)
-
-	blockID := base64.StdEncoding.EncodeToString([]byte(fmt.Sprintf("%6d", 0)))
-	_, err = bbClient.StageBlock(context.Background(), blockID, rsc, nil)
-	_require.Nil(err)
-
-	_, err = bbClient.CommitBlockList(context.Background(), []string{blockID}, nil)
-	_require.Nil(err)
-
-	resp, err := bbClient.GetBlockList(context.Background(), blockblob.BlockListTypeAll, nil)
-	_require.Nil(err)
-	_require.Len(resp.BlockList.CommittedBlocks, 1)
-	committed := resp.BlockList.CommittedBlocks
-	_require.Equal(*(committed[0].Name), blockID)
-	_require.Equal(*(committed[0].Size), largeBlockSize)
-	_require.Nil(resp.BlockList.UncommittedBlocks)
-}
-
 type fakeBlockBlob struct {
 	totalStaged int64
 }
@@ -3851,96 +3817,6 @@ func (f *fakeBlockBlob) Do(req *http.Request) (*http.Response, error) {
 		Header:     http.Header{},
 		Body:       http.NoBody,
 	}, nil
-}
-
-func TestUploadBufferUnevenBlockSize(t *testing.T) {
-	fbb := &fakeBlockBlob{}
-	client, err := blockblob.NewClientWithNoCredential("https://fake/blob/path", &blockblob.ClientOptions{
-		ClientOptions: policy.ClientOptions{
-			Transport: fbb,
-		},
-	})
-	require.NoError(t, err)
-	require.NotNil(t, client)
-
-	// create fake source that's not evenly divisible by 50000 (max number of blocks)
-	// and greater than MaxUploadBlobBytes (256MB) so that it doesn't fit into a single upload.
-
-	buffer := make([]byte, 263*1024*1024)
-	for i := 0; i < len(buffer); i++ {
-		buffer[i] = 1
-	}
-
-	_, err = client.UploadBuffer(context.Background(), buffer, &blockblob.UploadBufferOptions{
-		Concurrency: 1,
-	})
-	require.NoError(t, err)
-	require.Equal(t, int64(len(buffer)), fbb.totalStaged)
-}
-
-func TestUploadBufferEvenBlockSize(t *testing.T) {
-	fbb := &fakeBlockBlob{}
-	client, err := blockblob.NewClientWithNoCredential("https://fake/blob/path", &blockblob.ClientOptions{
-		ClientOptions: policy.ClientOptions{
-			Transport: fbb,
-		},
-	})
-	require.NoError(t, err)
-	require.NotNil(t, client)
-
-	// create fake source that's evenly divisible by 50000 (max number of blocks)
-	// and greater than MaxUploadBlobBytes (256MB) so that it doesn't fit into a single upload.
-
-	buffer := make([]byte, 270000000)
-	for i := 0; i < len(buffer); i++ {
-		buffer[i] = 1
-	}
-
-	_, err = client.UploadBuffer(context.Background(), buffer, &blockblob.UploadBufferOptions{
-		Concurrency: 1,
-	})
-	require.NoError(t, err)
-	require.Equal(t, int64(len(buffer)), fbb.totalStaged)
-}
-
-func TestUploadLogEvent(t *testing.T) {
-	listnercalled := false
-	log.SetEvents(azblob.EventUpload, log.EventRequest, log.EventResponse)
-	log.SetListener(func(cls log.Event, msg string) {
-		t.Logf("%s: %s\n", cls, msg)
-		if cls == azblob.EventUpload {
-			listnercalled = true
-			require.Equal(t, msg, "blob name path1/path2 actual size 270000000 block-size 4194304 block-count 65")
-		}
-	})
-
-	fbb := &fakeBlockBlob{}
-	client, err := blockblob.NewClientWithNoCredential("https://fake/blob/path1/path2", &blockblob.ClientOptions{
-		ClientOptions: policy.ClientOptions{
-			Transport: fbb,
-			Telemetry: policy.TelemetryOptions{ApplicationID: "testApp/1.0.0-preview.2"},
-		},
-	})
-	require.NoError(t, err)
-	require.NotNil(t, client)
-
-	// create fake source that's evenly divisible by 50000 (max number of blocks)
-	// and greater than MaxUploadBlobBytes (256MB) so that it doesn't fit into a single upload.
-
-	buffer := make([]byte, 270000000)
-	for i := 0; i < len(buffer); i++ {
-		buffer[i] = 1
-	}
-
-	_, err = client.UploadBuffer(context.Background(), buffer, &blockblob.UploadBufferOptions{
-		Concurrency: 1,
-		Progress: func(bytesTransferred int64) {
-			t.Logf("%v percent job done", (bytesTransferred*100)/270000000)
-		},
-	})
-	require.NoError(t, err)
-	require.Equal(t, int64(len(buffer)), fbb.totalStaged)
-	require.Equal(t, listnercalled, true)
 }
 
 type trequestIDPolicy struct{}
